requirements.txt
python-dotenv
flask==2.1.3
flask_sqlalchemy==3.1.1
flask-migrate
flask_validator
psycopg2-binary
Werkzeug==2.2.2